Added PC-student switch functionality and added SPC,temp1,temptables to database,so change the database

+++ IN/signup.php
@@ -7,44 +7,51 @@ if ($mysqli->connect_errno) {
     exit();
 }
 
+if(isset($passkey)) {
 
-$passkey=$_GET['passkey'];
-$tbl_name="temp";
+    $passkey = $_GET['passkey'];
+    $tbl_name = "temp";
 
 // Retrieve data from table where row that match this passkey 
-$sql="SELECT * FROM $tbl_name WHERE code ='$passkey'";
-$result=mysqli_query($mysqli,$sql);
+    $sql = "SELECT * FROM $tbl_name WHERE code ='$passkey'";
+    $result = mysqli_query($mysqli, $sql);
 
-if($result){
+    if ($result) {
 
 // Count how many row has this passkey
-$count=mysqli_num_rows($result);
+        $count = mysqli_num_rows($result);
 
 // if found this passkey in our database, retrieve data from table "temp_members_db"
-if($count==1){
+        if ($count == 1) {
 
-$rows=mysqli_fetch_array($result, MYSQLI_ASSOC);
-$name=$rows['name'];
-$email=$rows['email'];
-$usn=$rows['usn'];
+            $rows = mysqli_fetch_array($result, MYSQLI_ASSOC);
+            $name = $rows['name'];
+            $email = $rows['email'];
+            $usn = $rows['usn'];
 //$tbl_name2="";
 
 // Insert data that retrieves from "temp_members_db" into table "registered_members" 
 //$sql2="INSERT INTO $tbl_name2(name, email, password, country)VALUES('$name', '$email', '$password', '$country')";
 //$result2=mysql_query($sql2);
-/*
-// if successfully moved data from table"temp_members_db" to table "registered_members" displays message "Your account has been activated" and don't forget to delete confirmation code from table "temp_members_db"
-if($result2){
+            /*
+            // if successfully moved data from table"temp_members_db" to table "registered_members" displays message "Your account has been activated" and don't forget to delete confirmation code from table "temp_members_db"
+            if($result2){
 
-echo "Your account has been activated";
+            echo "Your account has been activated";
 
-// Delete information of this user from table "temp_members_db" that has this passkey 
-$sql3="DELETE FROM $tbl_name1 WHERE confirm_code = '$passkey'";
-$result3=mysql_query($sql3);
+            // Delete information of this user from table "temp_members_db" that has this passkey
+            $sql3="DELETE FROM $tbl_name1 WHERE confirm_code = '$passkey'";
+            $result3=mysql_query($sql3);
+
+            }
+            */
+        }
+    }
+}
+else {
+    echo "Wrong Confirmation code";
 
 }
-*/
-
 
 
 
@@ -243,13 +250,3 @@ $result3=mysql_query($sql3);
 </body>
 
 </html>
-
-<?php
-}
-
-// if not found passkey, display message "Wrong Confirmation code" 
-else {
-echo "Wrong Confirmation code";
-}
-}
-?>
